Fix an issue introduced by #15
When restoring the old geometry in fullscreen mode instead of readding, action must be taken, when the lower mode is very likely to change the `View`s geometry.

This is the case when the `Output` configuration changes and therefor we need to temporarily unset the fullscreened `View`, pass on the event, store the new geometry and reset fullscreen geometry.